solved

+++ src/test/java/com/pessimistic/aoc2024/util/PatternUtilsTest.java
@@ -5,38 +5,39 @@ import org.junit.jupiter.api.Test;
 
 import java.util.List;
 import java.util.stream.IntStream;
+import java.util.stream.LongStream;
 
 class PatternUtilsTest {
 
     @Test
     void isSameSignOrZero_true() {
-        Assertions.assertThat(PatternUtils.isSameSignOrZero(List.of(1, 2, 3, 4)))
+        Assertions.assertThat(PatternUtils.isSameSignOrZero(List.of(1L, 2L, 3L, 4L)))
                 .isTrue();
         Assertions.assertThat(PatternUtils.isSameSignOrZero(List.of()))
                 .isTrue();
-        Assertions.assertThat(PatternUtils.isSameSignOrZero(List.of(1, 0, 1, 0, 1)))
+        Assertions.assertThat(PatternUtils.isSameSignOrZero(List.of(1L, 0L, 1L, 0L, 1L)))
                 .isTrue();
-        Assertions.assertThat(PatternUtils.isSameSignOrZero(List.of(-111, -11, -1, 0)))
+        Assertions.assertThat(PatternUtils.isSameSignOrZero(List.of(-111L, -11L, -1L, 0L)))
                 .isTrue();
-        Assertions.assertThat(PatternUtils.isSameSignOrZero(List.of(0, 0)))
+        Assertions.assertThat(PatternUtils.isSameSignOrZero(List.of(0L, 0L)))
                 .isTrue();
     }
 
     @Test
     void isSameSignOrZero_false() {
-        Assertions.assertThat(PatternUtils.isSameSignOrZero(List.of(1, -2, 3, 4)))
+        Assertions.assertThat(PatternUtils.isSameSignOrZero(List.of(1L, -2L, 3L, 4L)))
                 .isFalse();
-        Assertions.assertThat(PatternUtils.isSameSignOrZero(List.of(1, 0, -1)))
+        Assertions.assertThat(PatternUtils.isSameSignOrZero(List.of(1L, 0L, -1L)))
                 .isFalse();
     }
 
     @Test
     void deltas() {
-        Assertions.assertThat(PatternUtils.deltas(IntStream.of(1, 0, -1)))
+        Assertions.assertThat(PatternUtils.deltas(LongStream.of(1L, 0L, -1L)))
                 .isEqualTo(List.of(-1, -1));
-        Assertions.assertThat(PatternUtils.deltas(IntStream.of(1, 2, 3, 4, 5)))
+        Assertions.assertThat(PatternUtils.deltas(LongStream.of(1L, 2L, 3L, 4L, 5L)))
                 .isEqualTo(List.of(1, 1, 1, 1));
-        Assertions.assertThat(PatternUtils.deltas(IntStream.of(10, 20, 30, 20, 10)))
+        Assertions.assertThat(PatternUtils.deltas(LongStream.of(10L, 20L, 30L, 20L, 10L)))
                 .isEqualTo(List.of(10, 10, -10, -10));
     }
 }
